internsynk/requirements.txt
asgiref==3.8.1
boto3==1.38.41
botocore==1.38.41
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
drf-yasg==1.21.10
environs==14.2.0
inflection==0.5.1
jmespath==1.0.1
marshmallow==4.0.0
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
PyMuPDF==1.26.1
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0